Electrometer/hhm based flyer
